+++ stoko_class_BLL/Product.cs
@@ -5,7 +5,7 @@ namespace stoko_class_BLL {
     public class Product {
         public int Id { get; set; }
         public String Title { get; set; }
-        public int PriceHT { get; set; }
+        public double PriceHT { get; set; }
         public String Reference { get; set; }
         private int quantity;
         public int Quantity {
@@ -32,7 +32,7 @@ namespace stoko_class_BLL {
         /// <param name="pQuantityStock"></param>
         /// <param name="pDescription"></param>
         /// <param name="pCategory"></param>
-        public Product(String pTitle, int pPriceHT, String pReference, int pQuantity, int pQuantityStock, String pDescription, Category pCategory = null) {
+        public Product(String pTitle, double pPriceHT, String pReference, int pQuantity, int pQuantityStock, String pDescription, Category pCategory = null) {
             Title = pTitle;
             PriceHT = pPriceHT;
             Reference = pReference;
@@ -53,7 +53,7 @@ namespace stoko_class_BLL {
         /// <param name="pQuantityStock"></param>
         /// <param name="pDescription"></param>
         /// <param name="pCategory"></param>
-        public Product(int pId, String pTitle, int pPriceHT, String pReference, int pQuantity, int pQuantityStock, String pDescription, Category pCategory = null) {
+        public Product(int pId, String pTitle, double pPriceHT, String pReference, int pQuantity, int pQuantityStock, String pDescription, Category pCategory = null) {
             Id = pId;
             Title = pTitle;
             PriceHT = pPriceHT;
